refactor(practice): replace target scoring with modal-based button scoring
Remove the buggy arrow-on-target placement flow and simplify the scoring
UX. Instead of two scoring methods (buttons vs target), scoring now uses
a single full-screen modal opened via a "Score now" button on each round
card. Also fixes the create payload to send `rounds` (not `ends`) and
serialize the date as an ISO string, matching the API validation schema.
