--- app/(tabs)/home/shooting-score.tsx
@@ -1,3 +1,0 @@
-import ShootingScoreScreen from '@/components/practice/screens/ShootingScoreScreen';
-
-export default ShootingScoreScreen;
