--- app/(tabs)/home/shooting.tsx
@@ -1,3 +1,0 @@
-import ShootingScreen from '@/components/practice/screens/ShootingScreen';
-
-export default ShootingScreen;
