+++ lib/i18n/types.ts
@@ -252,6 +252,10 @@ export interface TranslationKeys {
   'scoring.scoreSuffix': string;
   'scoring.editArrowAriaPrefix': string;
   'scoring.arrowScoring': string;
+  'scoring.scoreNow': string;
+  'scoring.editScores': string;
+  'scoring.continueScoring': string;
+  'scoring.done': string;
 
   // Reflection step UI
   'reflection.ratingLabel': string;
